backend/requirements.txt
aiohttp==3.8.6
fastapi==0.110.0
python-dotenv